[MERGE] merge of prognose branch hwv_sn_st_11_02 into svn head

+++ KalypsoCommons/src/org/kalypso/commons/java/io/FileUtilities.java
@@ -851,11 +851,14 @@ public class FileUtilities
   {
     try
     {
+      if( !file.exists() )
+        return;
+
       FileUtils.forceDelete( file );
     }
     catch( final IOException e )
     {
-      KalypsoCommonsPlugin.getDefault().getLog().log( StatusUtilities.statusFromThrowable( e ) );
+      KalypsoCommonsPlugin.getDefault().getLog().log( new Status( IStatus.ERROR, KalypsoCommonsPlugin.getID(), e.getLocalizedMessage(), e ) );
     }
   }
